+++ src/scripts/components/form/datepicker.vue
@@ -5,7 +5,7 @@
       :model="currentValue"
       :placeholder="placeholder"
       :plus="toggle || clear"
-      @change="handleInput">
+      @change="handleChange">
       <template slot="plus">
         <div v-if="toggle" class="mdl-datepicker__toggle" data-toggle>
           <slot name="toggle">
@@ -25,7 +25,11 @@
 <script>
 import Flatpickr from 'flatpickr';
 import UiTextfield from './textfield';
+import {isArray} from '../../helpers';
 
+const MODE_SINGLE = 'single';
+const MODE_MULTIPLE = 'multiple';
+const MODE_RANGE = 'range';
 const EVENT_CHANGE = 'change';
 
 export default {
@@ -56,32 +60,87 @@ export default {
   data() {
     return {
       flatpickr: null,
-      currentValue: this.model
-    }
-  },
-  methods: {
-    handleInput(event) {
-      this.currentValue = event.target.value;
-      this.$emit(EVENT_CHANGE, this.currentValue); // currentValue: string
+      currentValue: this.model,
+      mode: this.config.mode || MODE_SINGLE
     }
   },
   watch: {
     model(val) {
-      this.currentValue = val;
+      this.setRangeDate(val);
     }
   },
   mounted() {
     if (!this.flatpickr) {
+      // default config for ui
+      this.config.time_24hr = true;
       this.config.wrap = true;
+      // custom event
       this.config.onClose = () => {
         this.$refs.text.$el.querySelector('input').blur();
       };
+      // set default value
+      if (this.mode === MODE_SINGLE) {
+        this.config.onReady = (selectedDates, dateStr, instance) => {
+          // defaultDate: 'today'
+          if (dateStr) {
+            this.currentValue = dateStr;
+            this.$emit(EVENT_CHANGE, dateStr);
+          }
+        };
+      } else {
+        this.setRangeDate(this.model);
+        this.config.defaultDate = this.currentValue;
+      }
+      // init
       this.flatpickr = new Flatpickr(this.$el, this.config);
     }
   },
   beforeDestroy() {
     this.flatpickr.destroy();
     this.flatpickr = null;
+  },
+  methods: {
+    handleChange(event) {
+      this.currentValue = event.target.value;
+
+      let result;
+      switch (this.mode) {
+        case MODE_MULTIPLE:
+          let multipleValue = this.currentValue.replace(/\s,\s/, ',').split(',');
+          result = (multipleValue.length === 1)
+            ? multipleValue[0] // string
+            : multipleValue; // array
+          break;
+        case MODE_RANGE:
+          let rangeValue = this.currentValue.split(' to ');
+          let startDate = rangeValue[0];
+          let endDate = rangeValue[1];
+          if (startDate && endDate) {
+            result = (startDate === endDate)
+              ? startDate // string
+              : [startDate, endDate]; // array
+
+            this.flatpickr.setDate(result); // TODO 暂时不做实时同步显示(flatpickr bug)
+          }
+          break;
+        default:
+          result = this.currentValue; // string
+          break;
+      }
+
+      if (result) {
+        this.$emit(EVENT_CHANGE, result);
+      }
+    },
+    setRangeDate(value) {
+      if (this.config.mode === MODE_RANGE && isArray(this.value) && this.value.length === 2) {
+        let startDate = this.value[0];
+        let endDate = this.value[1];
+        this.currentValue = (startDate === endDate)
+          ? startDate
+          : `${startDate} to ${endDate}`;
+      }
+    }
   }
 };
 </script>